engine: extract esm/{sky_image_tile,tangent_tile}.js
