+++ engine/esm/transpiled.js
@@ -35,15 +35,11 @@ import { CAAMoon } from "./astrocalc/moon.js";
 import { AstroCalc } from "./astrocalc.js";
 
 import {
-  tileCacheAddTileToQueue,
-  tileCacheGetCachedTile,
   tileCacheGetTile,
-  tileCacheRemoveFromQueue,
-  tileDemEnabled,
   tilePrepDevice,
   tileUvMultiple,
+  useGl,
   useGlVersion2,
-  inc_tileCacheAccessID,
   set_tileCacheAddTileToQueue,
   set_tileCacheGetCachedTile,
   set_tileCacheGetTile,
@@ -51,6 +47,7 @@ import {
   set_tileDemEnabled,
   set_tilePrepDevice,
   set_tileUvMultiple,
+  set_useGl,
   set_useGlVersion2,
 } from "./render_globals.js";
 import { freestandingMode, set_freestandingMode } from "./data_globals.js";
@@ -72,7 +69,7 @@ import {
   Vector4d,
   PositionNormalTextured,
   SphereHull,
-  ConvexHull
+  ConvexHull,
 } from "./double3d.js";
 
 import { WEBGL } from "./graphics/webgl_constants.js";
@@ -94,7 +91,6 @@ import {
   EllipseShader,
   ModelShader,
   TileShader,
-  FitsShader,
   ImageShader,
   TextShader,
 } from "./graphics/shaders.js";
@@ -142,16 +138,17 @@ import { UiTools } from "./ui_tools.js";
 import { Star, Galaxy } from "./star.js";
 
 import { ColorMapContainer } from "./layers/color_map_container.js";
-import { WcsImage } from "./layers/wcs_image.js";
 import { FitsImage } from "./layers/fits_image.js";
 import { FitsImageJs } from "./layers/fits_image_js.js";
-import { FitsImageTile } from "./layers/fits_image_tile.js";
 
 import { RenderTriangle } from "./render_triangle.js";
+import { Tile } from "./tile.js";
 import { EquirectangularTile } from "./equirectangular_tile.js";
 import { HealpixTile } from "./healpix_tile.js";
 import { MercatorTile } from "./mercator_tile.js";
 import { PlotTile } from "./plot_tile.js";
+import { TangentTile } from "./tangent_tile.js";
+import { SkyImageTile } from "./sky_image_tile.js";
 
 
 // wwtlib.ScaleTypes
@@ -11649,7 +11646,6 @@ export function RenderContext() {
   }
 }
 
-RenderContext.useGl = false;
 RenderContext.back = 0;
 
 RenderContext.create = function (device) {
@@ -14454,778 +14450,6 @@ var SpaceTimeController$ = {};
 
 registerType("SpaceTimeController", [SpaceTimeController, SpaceTimeController$, null]);
 
-// wwtlib.LatLngEdges
-
-export function LatLngEdges() {
-  this.latMin = 0;
-  this.latMax = 0;
-  this.lngMin = 0;
-  this.lngMax = 0;
-}
-
-var LatLngEdges$ = {};
-
-registerType("LatLngEdges", [LatLngEdges, LatLngEdges$, null]);
-
-// wwtlib.Tile
-
-export function Tile() {
-  this._renderTriangleLists = new Array(4);
-  this._indexBuffers = new Array(4);
-  this.level = 0;
-  this.tileX = 0;
-  this.tileY = 0;
-  this.texture = null;
-  this.texture2d = null;
-  this.isCatalogTile = false;
-  this.readyToRender = false;
-  this.inViewFrustum = true;
-  this.globalCenter = Vector3d.zero;
-  this.children = [null, null, null, null];
-  this.parent = null;
-  this.localCenter = new Vector3d();
-  this.renderedAtOrBelowGeneration = 0;
-  this._demScaleFactor = 6371000;
-  this.demIndex = 0;
-  this.demAverage = 0;
-  this.demReady = false;
-  this.texReady = false;
-  this.demTile = false;
-  this.demDownloading = false;
-  this.renderedGeneration = 0;
-  this.accomidation = 0;
-  this.accessCount = 0;
-  this.downloading = false;
-  this.geometryCreated = false;
-  this._isHdTile = false;
-  this.demSize = 33 * 33;
-  this._topLeftScreen = new Vector3d();
-  this._bottomRightScreen = new Vector3d();
-  this._topRightScreen = new Vector3d();
-  this._bottomLeftScreen = new Vector3d();
-  this.sphereRadius = 0;
-  this.sphereCenter = new Vector3d();
-  this.radius = 1;
-  this.triangleCount = 0;
-  this.requestHits = 0;
-  this.requestPending = false;
-  this.errored = false;
-  this._key = null;
-  this._tileId = null;
-  this._vertexCount = 0;
-  this.renderChildPart = null;
-  this.renderChildPart = new Array(4);
-  for (var i = 0; i < 4; i++) {
-    this.renderChildPart[i] = BlendState.create(false, 500);
-  }
-}
-
-Tile.currentRenderGeneration = 0;
-Tile.tileTargetX = -1;
-Tile.tileTargetY = -1;
-Tile.tileTargetLevel = -1;
-Tile.tilesInView = 0;
-Tile.trianglesRendered = 0;
-Tile.tilesTouched = 0;
-Tile.frustumList = null;
-set_tileUvMultiple(1);
-Tile.callCount = 0;
-Tile.useAccomidation = true;
-set_tileDemEnabled(true);
-Tile.maxLevel = 20;
-Tile.meshComplexity = 50;
-Tile.imageQuality = 50;
-Tile.lastDeepestLevel = 0;
-Tile.deepestLevel = 0;
-Tile.RC = (3.1415927 / 180);
-
-Tile.getFrustumList = function () {
-  try {
-    return Tile.frustumList;
-  }
-  catch ($e1) {
-    return null;
-  }
-};
-
-Tile.get_subDivisions = function () {
-  return 32;
-};
-
-var Tile$ = {
-  getIndexBuffer: function (index, accomidation) {
-    return this._indexBuffers[index];
-  },
-
-  isPointInTile: function (lat, lng) {
-    return false;
-  },
-
-  getSurfacePointAltitude: function (lat, lng, meters) {
-    return 0;
-  },
-
-  makeTexture: function () {
-    if (tilePrepDevice != null) {
-      try {
-        this.texture2d = tilePrepDevice.createTexture();
-        tilePrepDevice.bindTexture(WEBGL.TEXTURE_2D, this.texture2d);
-        tilePrepDevice.texParameteri(WEBGL.TEXTURE_2D, WEBGL.TEXTURE_WRAP_S, WEBGL.CLAMP_TO_EDGE);
-        tilePrepDevice.texParameteri(WEBGL.TEXTURE_2D, WEBGL.TEXTURE_WRAP_T, WEBGL.CLAMP_TO_EDGE);
-        if (this.dataset.get_extension().toLowerCase().indexOf('fits') > -1 && useGlVersion2) {
-          tilePrepDevice.texImage2D(WEBGL.TEXTURE_2D, 0, WEBGL.R32F, ss.truncate(this.fitsImage.get_sizeX()), ss.truncate(this.fitsImage.get_sizeY()), 0, WEBGL.RED, WEBGL.FLOAT, this.fitsImage.dataUnit);
-          tilePrepDevice.texParameteri(WEBGL.TEXTURE_2D, WEBGL.TEXTURE_MIN_FILTER, WEBGL.NEAREST);
-          tilePrepDevice.texParameteri(WEBGL.TEXTURE_2D, WEBGL.TEXTURE_MAG_FILTER, WEBGL.NEAREST);
-        }
-        else {
-          var image = this.texture;
-          if ((!Texture.isPowerOfTwo(this.texture.height) | !Texture.isPowerOfTwo(this.texture.width)) === 1) {
-            var temp = document.createElement('canvas');
-            temp.height = Texture.fitPowerOfTwo(image.height);
-            temp.width = Texture.fitPowerOfTwo(image.width);
-            var ctx = temp.getContext('2d');
-            ctx.drawImage(image, 0, 0, temp.width, temp.height);
-            image = temp;
-          }
-          tilePrepDevice.texImage2D(WEBGL.TEXTURE_2D, 0, WEBGL.RGBA, WEBGL.RGBA, WEBGL.UNSIGNED_BYTE, image);
-          tilePrepDevice.texParameteri(WEBGL.TEXTURE_2D, WEBGL.TEXTURE_MIN_FILTER, WEBGL.LINEAR_MIPMAP_NEAREST);
-          tilePrepDevice.generateMipmap(WEBGL.TEXTURE_2D);
-        }
-        tilePrepDevice.bindTexture(WEBGL.TEXTURE_2D, null);
-      }
-      catch ($e1) {
-        this.errored = true;
-      }
-    }
-  },
-
-  addVertex: function (buffer, index, p) {
-    buffer[index++] = p.position.x;
-    buffer[index++] = p.position.y;
-    buffer[index++] = p.position.z;
-    buffer[index++] = p.tu;
-    buffer[index++] = p.tv;
-    return index;
-  },
-
-  geoTo3dWithAlt: function (lat, lng, useLocalCenter, rev) {
-    lat = Math.max(Math.min(90, lat), -90);
-    lng = Math.max(Math.min(180, lng), -180);
-    if (!tileDemEnabled || this.demData == null) {
-      return this.geoTo3d(lat, lng, useLocalCenter);
-    }
-    if (rev) {
-      lng -= 180;
-    }
-    var altitude = this.demData[this.demIndex];
-    var retVal = this.geoTo3dWithAltitude(lat, lng, altitude, useLocalCenter);
-    return retVal;
-  },
-
-  geoTo3dWithAltitude: function (lat, lng, altitude, useLocalCenter) {
-    var radius = 1 + (altitude / this.get__demScaleFactor());
-    var retVal = Vector3d.create((Math.cos(lng * Tile.RC) * Math.cos(lat * Tile.RC) * radius), (Math.sin(lat * Tile.RC) * radius), (Math.sin(lng * Tile.RC) * Math.cos(lat * Tile.RC) * radius));
-    if (useLocalCenter) {
-      retVal.subtract(this.localCenter);
-    }
-    return retVal;
-  },
-
-  get__demScaleFactor: function () {
-    return this._demScaleFactor;
-  },
-
-  set__demScaleFactor: function (value) {
-    this._demScaleFactor = value;
-    return value;
-  },
-
-  requestImage: function () {
-    var $this = this;
-
-    if (this.dataset.get_extension().toLowerCase().indexOf('fits') > -1) {
-      if (!this.downloading && !this.readyToRender) {
-        this.downloading = true;
-        if (useGlVersion2) {
-          this.fitsImage = new FitsImageTile(this.dataset, this.get_URL(), function (wcsImage) {
-            $this.downloading = false;
-            $this.errored = $this.fitsImage.errored;
-            tileCacheRemoveFromQueue($this.get_key(), true);
-            if (!$this.fitsImage.errored) {
-              if (!$this.level) {
-                $this.dataset.get_fitsProperties()._fireMainImageLoaded($this.fitsImage);
-                $this.fitsImage.applyDisplaySettings();
-              }
-              $this.texReady = true;
-              $this.readyToRender = $this.texReady && ($this.demReady || !$this.demTile);
-              $this.requestPending = false;
-              $this.makeTexture();
-            }
-          });
-        }
-        else {
-          this.fitsImage = FitsImageJs.createTiledFits(this.dataset, this.get_URL(), function (wcsImage) {
-            if (!$this.level) {
-              $this.dataset.get_fitsProperties()._fireMainImageLoaded($this.fitsImage);
-            }
-            $this.texReady = true;
-            $this.downloading = false;
-            $this.errored = $this.fitsImage.errored;
-            $this.readyToRender = $this.texReady && ($this.demReady || !$this.demTile);
-            $this.requestPending = false;
-            tileCacheRemoveFromQueue($this.get_key(), true);
-            $this.texture2d = wcsImage.getBitmap().getTexture();
-          });
-        }
-      }
-    } else {
-      if (this.get_dataset().get_wcsImage() != null) {
-        this.texReady = true;
-        this.downloading = false;
-        this.errored = false;
-        this.readyToRender = true;
-        this.requestPending = false;
-        tileCacheRemoveFromQueue(this.get_key(), true);
-        return;
-      }
-      if (!this.downloading && !this.readyToRender) {
-        this.downloading = true;
-        this.texture = document.createElement('img');
-        var xdomimg = this.texture;
-        this.texture.addEventListener('load', function (e) {
-          $this.texReady = true;
-          $this.downloading = false;
-          $this.errored = false;
-          $this.readyToRender = $this.texReady && ($this.demReady || !$this.demTile);
-          $this.requestPending = false;
-          tileCacheRemoveFromQueue($this.get_key(), true);
-          $this.makeTexture();
-        }, false);
-        this.texture.addEventListener('error', function (e) {
-          if (!$this.texture.hasAttribute('proxyattempt')) {
-            $this.texture.setAttribute('proxyattempt', true);
-            var new_url = URLHelpers.singleton.activateProxy($this.texture.src);
-            if (new_url != null) {
-              $this.texture.src = new_url;
-              return;
-            }
-          }
-          $this.downloading = false;
-          $this.readyToRender = false;
-          $this.errored = true;
-          $this.requestPending = false;
-          tileCacheRemoveFromQueue($this.get_key(), true);
-        }, false);
-        xdomimg.crossOrigin = 'anonymous';
-        this.texture.src = this.get_URL();
-      }
-    }
-  },
-
-  createDemFromParent: function () {
-    return false;
-  },
-
-  _loadDemData: function () {
-    if (this.demFile == null) {
-      return this.createDemFromParent();
-    }
-    this.demData = this.demFile;
-    if (this.demFile.length !== 1089 && this.demFile.length !== 513) {
-      return this.createDemFromParent();
-    }
-    var total = 0;
-    var $enum1 = ss.enumerate(this.demData);
-    while ($enum1.moveNext()) {
-      var fv = $enum1.current;
-      total += fv;
-    }
-    this.demAverage /= this.demData.length;
-    return true;
-  },
-
-  requestDem: function () {
-    var $this = this;
-
-    if (!this.readyToRender && !this.demDownloading) {
-      this.demTile = true;
-      this.demDownloading = true;
-      Tile.callCount++;
-      var xhr = new XMLHttpRequest();
-      xhr.addEventListener('load', function (e) {
-        $this.demReady = true;
-        $this.demDownloading = false;
-        $this.readyToRender = $this.texReady && ($this.demReady || !$this.demTile);
-        $this.requestPending = false;
-        try {
-          $this.demFile = new Float32Array(xhr.response);
-        }
-        catch ($e1) {
-        }
-        tileCacheRemoveFromQueue($this.get_key(), true);
-      }, false);
-      xhr.addEventListener('error', function (e) {
-        $this.demDownloading = false;
-        $this.demReady = false;
-        $this.readyToRender = false;
-        $this.errored = true;
-        $this.requestPending = false;
-        tileCacheRemoveFromQueue($this.get_key(), true);
-      }, false);
-      xhr.open('GET', this.get_demURL(), true);
-      xhr.responseType = 'arraybuffer';
-      xhr.send();
-    }
-  },
-
-  draw3D: function (renderContext, opacity) {
-    this.renderedGeneration = Tile.currentRenderGeneration;
-    Tile.tilesTouched++;
-    this.accessCount = inc_tileCacheAccessID();
-    if (this.errored) {
-      return false;
-    }
-    var xMax = 2;
-    this.inViewFrustum = true;
-    if (!this.readyToRender) {
-      tileCacheAddTileToQueue(this);
-      return false;
-    }
-    var transitioning = false;
-    var childIndex = 0;
-    var yOffset = 0;
-    if (this.dataset.get_mercator() || this.dataset.get_bottomsUp()) {
-      yOffset = 1;
-    }
-    var xOffset = 0;
-    var anythingToRender = false;
-    var childRendered = false;
-    for (var y1 = 0; y1 < 2; y1++) {
-      for (var x1 = 0; x1 < xMax; x1++) {
-        if (this.level < this.dataset.get_levels()) {
-          if (this.children[childIndex] == null) {
-            this.children[childIndex] = tileCacheGetTile(this.level + 1, this.tileX * 2 + ((x1 + xOffset) % 2), this.tileY * 2 + ((y1 + yOffset) % 2), this.dataset, this);
-          }
-          if (this.children[childIndex].isTileInFrustum(renderContext.get_frustum())) {
-            this.inViewFrustum = true;
-            if (this.children[childIndex].isTileBigEnough(renderContext)) {
-              this.renderChildPart[childIndex].set_targetState(!this.children[childIndex].draw3D(renderContext, opacity));
-              if (this.renderChildPart[childIndex].get_targetState()) {
-                childRendered = true;
-              }
-            }
-            else {
-              this.renderChildPart[childIndex].set_targetState(true);
-            }
-          }
-          else {
-            this.renderChildPart[childIndex].set_targetState(this.renderChildPart[childIndex].set_state(false));
-          }
-          if (this.renderChildPart[childIndex].get_targetState() !== this.renderChildPart[childIndex].get_state()) {
-            transitioning = true;
-          }
-        }
-        else {
-          this.renderChildPart[childIndex].set_state(true);
-        }
-        if (!!this.renderChildPart[childIndex].get_state()) {
-          anythingToRender = true;
-        }
-        childIndex++;
-      }
-    }
-    if (childRendered || anythingToRender) {
-      this.renderedAtOrBelowGeneration = Tile.currentRenderGeneration;
-      if (this.parent != null) {
-        this.parent.renderedAtOrBelowGeneration = this.renderedAtOrBelowGeneration;
-      }
-    }
-    if (!anythingToRender) {
-      return true;
-    }
-    if (!this.createGeometry(renderContext)) {
-      return false;
-    }
-    Tile.tilesInView++;
-    this.accomidation = this._computeAccomidation();
-    for (var i = 0; i < 4; i++) {
-      if (this.renderChildPart[i].get_targetState()) {
-        this.renderPart(renderContext, i, (opacity / 100), false);
-      }
-    }
-    return true;
-  },
-
-  _computeAccomidation: function () {
-    var accVal = 0;
-    if (!Tile.useAccomidation) {
-      return 0;
-    }
-    var top = tileCacheGetCachedTile(this.level, this.tileX, this.tileY + 1, this.dataset, this);
-    if (top == null || top.renderedAtOrBelowGeneration < Tile.currentRenderGeneration - 2) {
-      accVal += 1;
-    }
-    var right = tileCacheGetCachedTile(this.level, this.tileX + 1, this.tileY, this.dataset, this);
-    if (right == null || right.renderedAtOrBelowGeneration < Tile.currentRenderGeneration - 2) {
-      accVal += 2;
-    }
-    var bottom = tileCacheGetCachedTile(this.level, this.tileX, this.tileY - 1, this.dataset, this);
-    if (bottom == null || bottom.renderedAtOrBelowGeneration < Tile.currentRenderGeneration - 2) {
-      accVal += 4;
-    }
-    var left = tileCacheGetCachedTile(this.level, this.tileX - 1, this.tileY, this.dataset, this);
-    if (left == null || left.renderedAtOrBelowGeneration < Tile.currentRenderGeneration - 2) {
-      accVal += 8;
-    }
-    return accVal;
-  },
-
-  renderPart: function (renderContext, part, opacity, combine) {
-    if (tilePrepDevice == null) {
-      var lighting = renderContext.lighting && renderContext.get_sunPosition() != null;
-      var $enum1 = ss.enumerate(this._renderTriangleLists[part]);
-      while ($enum1.moveNext()) {
-        var tri = $enum1.current;
-        tri.opacity = opacity;
-        if (lighting) {
-          var norm = tri.normal.copy();
-          renderContext.get_world().multiplyVector(norm);
-          norm.normalize();
-          var light = Vector3d.dot(norm, renderContext.get_sunPosition());
-          if (light < 0) {
-            light = 0;
-          }
-          else {
-            light = Math.min(1, (light * 1));
-          }
-          tri.lighting = light;
-        }
-        else {
-          tri.lighting = 1;
-        }
-        tri.draw(renderContext.device, renderContext.WVP);
-      }
-    } else {
-      if (useGlVersion2 && this.fitsImage != null) {
-        ColorMapContainer.bindColorMapTexture(tilePrepDevice, this.dataset.get_fitsProperties().colorMapName);
-        FitsShader.min = this.dataset.get_fitsProperties().lowerCut;
-        FitsShader.max = this.dataset.get_fitsProperties().upperCut;
-        FitsShader.containsBlanks = this.dataset.get_fitsProperties().containsBlanks;
-        FitsShader.blankValue = this.dataset.get_fitsProperties().blankValue;
-        FitsShader.bZero = this.dataset.get_fitsProperties().bZero;
-        FitsShader.bScale = this.dataset.get_fitsProperties().bScale;
-        FitsShader.scaleType = this.dataset.get_fitsProperties().scaleType;
-        FitsShader.transparentBlack = this.dataset.get_fitsProperties().transparentBlack;
-        FitsShader.use(renderContext, this._vertexBuffer, this.getIndexBuffer(part, this.accomidation), this.texture2d, opacity, false, this.globalCenter);
-      }
-      else {
-        TileShader.use(renderContext, this._vertexBuffer, this.getIndexBuffer(part, this.accomidation), this.texture2d, opacity, false, this.globalCenter);
-      }
-      renderContext.gl.drawElements(WEBGL.TRIANGLES, this.triangleCount * 3, WEBGL.UNSIGNED_SHORT, 0);
-    }
-  },
-
-  cleanUp: function (removeFromParent) {
-    this.readyToRender = false;
-    this.demData = null;
-    this.demFile = null;
-    this.demDownloading = false;
-    this.texReady = false;
-    this.demReady = false;
-    this.errored = false;
-    if (this.texture != null) {
-      this.texture = null;
-    }
-    this._renderTriangleLists = new Array(4);
-    this.geometryCreated = false;
-    if (removeFromParent && this.parent != null) {
-      this.parent.removeChild(this);
-      this.parent = null;
-    }
-    if (tilePrepDevice != null) {
-      var $enum1 = ss.enumerate(this._indexBuffers);
-      while ($enum1.moveNext()) {
-        var buf = $enum1.current;
-        tilePrepDevice.deleteBuffer(buf);
-      }
-      this._indexBuffers = new Array(4);
-      if (this._vertexBuffer != null) {
-        tilePrepDevice.deleteBuffer(this._vertexBuffer);
-        this._vertexBuffer = null;
-      }
-      if (this.texture2d != null) {
-        tilePrepDevice.deleteTexture(this.texture2d);
-        this.texture2d = null;
-      }
-    }
-  },
-
-  removeChild: function (child) {
-    for (var i = 0; i < 4; i++) {
-      if (this.children[i] === child) {
-        this.children[i] = null;
-        return;
-      }
-    }
-  },
-
-  createGeometry: function (renderContext) {
-    if (tileDemEnabled && this.demReady && this.demData == null) {
-      if (!this._loadDemData()) {
-        return false;
-      }
-    }
-    if (tileDemEnabled && this.demData == null) {
-      return false;
-    }
-    this.readyToRender = true;
-    return true;
-  },
-
-  calcSphere: function () {
-    var corners = new Array(4);
-    corners[0] = this.topLeft;
-    corners[1] = this.bottomRight;
-    corners[2] = this.topRight;
-    corners[3] = this.bottomLeft;
-    var result = ConvexHull.findEnclosingSphere(corners);
-    this.sphereCenter = result.center;
-    this.sphereRadius = result.radius;
-  },
-
-  isTileBigEnough: function (renderContext) {
-    if (this.level > 1) {
-      var wvp = renderContext.WVP;
-      wvp._transformTo(this.topLeft, this._topLeftScreen);
-      wvp._transformTo(this.bottomRight, this._bottomRightScreen);
-      wvp._transformTo(this.topRight, this._topRightScreen);
-      wvp._transformTo(this.bottomLeft, this._bottomLeftScreen);
-      var top = this._topLeftScreen;
-      top.subtract(this._topRightScreen);
-      var topLength = top.length();
-      var bottom = this._bottomLeftScreen;
-      bottom.subtract(this._bottomRightScreen);
-      var bottomLength = bottom.length();
-      var left = this._bottomLeftScreen;
-      left.subtract(this._topLeftScreen);
-      var leftLength = left.length();
-      var right = this._bottomRightScreen;
-      right.subtract(this._topRightScreen);
-      var rightLength = right.length();
-      var lengthMax = Math.max(Math.max(rightLength, leftLength), Math.max(bottomLength, topLength));
-      if (lengthMax < 300) {
-        return false;
-      }
-      else {
-        Tile.deepestLevel = (this.level > Tile.deepestLevel) ? this.level : Tile.deepestLevel;
-      }
-    }
-    return true;
-  },
-
-  isTileInFrustum: function (frustum) {
-    if (this.level < 2 && (!this.dataset.get_projection() || this.dataset.get_projection() === 3)) {
-    }
-    this.inViewFrustum = false;
-    var centerV4 = new Vector4d(this.sphereCenter.x, this.sphereCenter.y, this.sphereCenter.z, 1);
-    for (var i = 0; i < 6; i++) {
-      if (frustum[i].dot(centerV4) < -this.sphereRadius) {
-        return false;
-      }
-    }
-    this.inViewFrustum = true;
-    return true;
-  },
-
-  get_sphereRadius: function () {
-    return this.sphereRadius;
-  },
-
-  get_sphereCenter: function () {
-    return this.sphereCenter;
-  },
-
-  geoTo3d: function (lat, lng, useLocalCenter) {
-    if (this.dataset.get_dataSetType() === 3) {
-      var retVal = Vector3d.create(-(Math.cos(lng * Tile.RC) * Math.cos(lat * Tile.RC) * this.radius), (Math.sin(lat * Tile.RC) * this.radius), (Math.sin(lng * Tile.RC) * Math.cos(lat * Tile.RC) * this.radius));
-      return retVal;
-    } else {
-      lng -= 180;
-      var retVal = Vector3d.create((Math.cos(lng * Tile.RC) * Math.cos(lat * Tile.RC) * this.radius), (Math.sin(lat * Tile.RC) * this.radius), (Math.sin(lng * Tile.RC) * Math.cos(lat * Tile.RC) * this.radius));
-      return retVal;
-    }
-  },
-
-  onCreateVertexBuffer: function (sender, e) { },
-
-  get_dataset: function () {
-    return this.dataset;
-  },
-
-  set_dataset: function (value) {
-    this.dataset = value;
-    return value;
-  },
-
-  get_key: function () {
-    if (this._key == null) {
-      this._key = getTileKey(this.dataset, this.level, this.tileX, this.tileY, this.parent);
-    }
-    return this._key;
-  },
-
-  get_URL: function () {
-    var rewritten_url = URLHelpers.singleton.rewrite(this.dataset.get_url(), 0);
-    var returnUrl = rewritten_url;
-    if (rewritten_url.indexOf('{1}') > -1) {
-      if (!this.dataset.get_projection() && !ss.emptyString(this.dataset.get_quadTreeTileMap())) {
-        returnUrl = ss.format(rewritten_url, this.getServerID(), this.getTileID());
-        if (returnUrl.indexOf('virtualearth.net') > -1) {
-          returnUrl += '&n=z';
-        }
-        return returnUrl;
-      }
-      else {
-        return ss.format(rewritten_url, this.dataset.get_imageSetID(), this.level, this.tileX, this.tileY);
-      }
-    }
-    returnUrl = ss.replaceString(returnUrl, '{X}', this.tileX.toString());
-    returnUrl = ss.replaceString(returnUrl, '{Y}', this.tileY.toString());
-    returnUrl = ss.replaceString(returnUrl, '{L}', this.level.toString());
-    var hash = 0;
-    if (returnUrl.indexOf('{S:0}') > -1) {
-      hash = 0;
-      returnUrl = ss.replaceString(returnUrl, '{S:0}', '{S}');
-    }
-    if (returnUrl.indexOf('{S:1}') > -1) {
-      hash = 1;
-      returnUrl = ss.replaceString(returnUrl, '{S:1}', '{S}');
-    }
-    if (returnUrl.indexOf('{S:2}') > -1) {
-      hash = 2;
-      returnUrl = ss.replaceString(returnUrl, '{S:2}', '{S}');
-    }
-    if (returnUrl.indexOf('{S:3}') > -1) {
-      hash = 3;
-      returnUrl = ss.replaceString(returnUrl, '{S:3}', '{S}');
-    }
-    if (returnUrl.indexOf('a{S}') > -1) {
-      returnUrl = ss.replaceString(returnUrl, 'a{S}', 'r{S}');
-    }
-    if (returnUrl.indexOf('h{S}') > -1) {
-      returnUrl = ss.replaceString(returnUrl, 'h{S}', 'r{S}');
-    }
-    if (returnUrl.indexOf('//r{S}.ortho.tiles.virtualearth.net') > -1) {
-      returnUrl = ss.replaceString(returnUrl, '//r{S}.ortho.tiles.virtualearth.net', '//ecn.t{S}.tiles.virtualearth.net');
-    }
-    var id = this.getTileID();
-    var server = '';
-    if (!id.length) {
-      server = hash.toString();
-    } else {
-      server = id.substr(id.length - 1, 1);
-    }
-    returnUrl = ss.replaceString(returnUrl, '{Q}', id);
-    returnUrl = ss.replaceString(returnUrl, '{S}', server);
-    if (returnUrl.indexOf('virtualearth.net') > -1) {
-      returnUrl += '&n=z';
-    }
-    return returnUrl;
-  },
-
-  get_demURL: function () {
-    var rewritten_url = URLHelpers.singleton.rewrite(this.dataset.get_demUrl(), 0);
-    if (!this.dataset.get_projection() && !freestandingMode) {
-      var baseUrl = URLHelpers.singleton.coreStaticUrl('wwtweb/demtile.aspx?q={0},{1},{2},M');
-      if (!ss.emptyString(rewritten_url)) {
-        baseUrl = rewritten_url;
-      }
-    }
-    if (rewritten_url.indexOf('{1}') > -1) {
-      return ss.format(rewritten_url + '&new', this.level, this.tileX, this.tileY);
-    }
-    var returnUrl = rewritten_url;
-    returnUrl = ss.replaceString(returnUrl, '{X}', this.tileX.toString());
-    returnUrl = ss.replaceString(returnUrl, '{Y}', this.tileY.toString());
-    returnUrl = ss.replaceString(returnUrl, '{L}', this.level.toString());
-    var hash = 0;
-    if (returnUrl.indexOf('{S:0}') > -1) {
-      hash = 0;
-      returnUrl = ss.replaceString(returnUrl, '{S:0}', '{S}');
-    }
-    if (returnUrl.indexOf('{S:1}') > -1) {
-      hash = 1;
-      returnUrl = ss.replaceString(returnUrl, '{S:1}', '{S}');
-    }
-    if (returnUrl.indexOf('{S:2}') > -1) {
-      hash = 2;
-      returnUrl = ss.replaceString(returnUrl, '{S:2}', '{S}');
-    }
-    if (returnUrl.indexOf('{S:3}') > -1) {
-      hash = 3;
-      returnUrl = ss.replaceString(returnUrl, '{S:3}', '{S}');
-    }
-    var id = this.getTileID();
-    var server = '';
-    if (!id.length) {
-      server = hash.toString();
-    } else {
-      server = id.substr(id.length - 1, 1);
-    }
-    returnUrl = ss.replaceString(returnUrl, '{Q}', id);
-    returnUrl = ss.replaceString(returnUrl, '{S}', server);
-    return returnUrl;
-  },
-
-  getServerID: function () {
-    var server = (this.tileX & 1) + ((this.tileY & 1) << 1);
-    return server;
-  },
-
-  getTileID: function () {
-    if (this._tileId != null) {
-      return this._tileId;
-    }
-    var netLevel = this.level;
-    var netX = this.tileX;
-    var netY = this.tileY;
-    if (this.dataset.get_projection() === 1) {
-      netLevel++;
-    }
-    var tileMap = this.dataset.get_quadTreeTileMap();
-    if (!ss.emptyString(tileMap)) {
-      var sb = new ss.StringBuilder();
-      for (var i = netLevel; i > 0; --i) {
-        var mask = 1 << (i - 1);
-        var val = 0;
-        if (!!(netX & mask)) {
-          val = 1;
-        }
-        if (!!(netY & mask)) {
-          val += 2;
-        }
-        sb.append(tileMap.substr(val, 1));
-      }
-      this._tileId = sb.toString();
-      return this._tileId;
-    } else {
-      this._tileId = '0';
-      return this._tileId;
-    }
-  },
-
-  get_vertexCount: function () {
-    return this._vertexCount;
-  },
-
-  set_vertexCount: function (value) {
-    this._vertexCount = value;
-    return value;
-  }
-};
-
-registerType("Tile", [Tile, Tile$, null]);
-
 // wwtlib.TileCache
 
 export function TileCache() { }
@@ -16010,7 +15234,7 @@ var Overlay$ = {
   },
 
   draw3D: function (renderContext, designTime) {
-    if (RenderContext.useGl) {
+    if (useGl) {
       if (this.texture == null || this.isDynamic) {
         this.initializeTexture();
       }
@@ -23087,7 +22311,7 @@ WWTControl.initControl6 = function (DivId, startRenderLoop, startLat, startLng, 
     } else {
       set_tilePrepDevice(gl);
       WWTControl.singleton.renderContext.gl = gl;
-      RenderContext.useGl = true;
+      set_useGl(true);
     }
     WWTControl.singleton.canvas = canvas;
     WWTControl.singleton.renderContext.width = canvas.width;
@@ -30128,189 +29352,6 @@ var VoTableLayer$ = {
 
 registerType("VoTableLayer", [VoTableLayer, VoTableLayer$, Layer]);
 
-// wwtlib.TangentTile
-
-export function TangentTile(level, x, y, dataset, parent) {
-  this._topDown$1 = true;
-  Tile.call(this);
-  this.parent = parent;
-  this.level = level;
-  this.tileX = x;
-  this.tileY = y;
-  this.dataset = dataset;
-  this._topDown$1 = !dataset.get_bottomsUp();
-  this.computeBoundingSphere();
-}
-
-var TangentTile$ = {
-  computeBoundingSphere: function () {
-    if (!this._topDown$1) {
-      this.computeBoundingSphereBottomsUp();
-      return;
-    }
-    var tileDegrees = this.dataset.get_baseTileDegrees() / Math.pow(2, this.level);
-    var latMin = (this.dataset.get_baseTileDegrees() / 2 - (this.tileY) * tileDegrees) + this.dataset.get_offsetY();
-    var latMax = (this.dataset.get_baseTileDegrees() / 2 - ((this.tileY + 1)) * tileDegrees) + this.dataset.get_offsetY();
-    var lngMin = ((this.tileX) * tileDegrees - this.dataset.get_baseTileDegrees() / this.dataset.get_widthFactor()) + this.dataset.get_offsetX();
-    var lngMax = (((this.tileX + 1)) * tileDegrees - this.dataset.get_baseTileDegrees() / this.dataset.get_widthFactor()) + this.dataset.get_offsetX();
-    var latCenter = (latMin + latMax) / 2;
-    var lngCenter = (lngMin + lngMax) / 2;
-    this.sphereCenter = this.geoTo3dTan(latCenter, lngCenter);
-    this.topLeft = this.geoTo3dTan(latMin, lngMin);
-    this.bottomRight = this.geoTo3dTan(latMax, lngMax);
-    this.topRight = this.geoTo3dTan(latMin, lngMax);
-    this.bottomLeft = this.geoTo3dTan(latMax, lngMin);
-    var distVect = this.geoTo3dTan(latMin, lngMin);
-    distVect.subtract(this.sphereCenter);
-    this.sphereRadius = distVect.length();
-  },
-
-  computeBoundingSphereBottomsUp: function () {
-    var tileDegrees = this.dataset.get_baseTileDegrees() / Math.pow(2, this.level);
-    var latMin = (this.dataset.get_baseTileDegrees() / 2 + ((this.tileY + 1)) * tileDegrees) + this.dataset.get_offsetY();
-    var latMax = (this.dataset.get_baseTileDegrees() / 2 + (this.tileY) * tileDegrees) + this.dataset.get_offsetY();
-    var lngMin = ((this.tileX) * tileDegrees - this.dataset.get_baseTileDegrees() / this.dataset.get_widthFactor()) + this.dataset.get_offsetX();
-    var lngMax = (((this.tileX + 1)) * tileDegrees - this.dataset.get_baseTileDegrees() / this.dataset.get_widthFactor()) + this.dataset.get_offsetX();
-    this.topLeft = this.geoTo3dTan(latMin, lngMin);
-    this.bottomRight = this.geoTo3dTan(latMax, lngMax);
-    this.topRight = this.geoTo3dTan(latMin, lngMax);
-    this.bottomLeft = this.geoTo3dTan(latMax, lngMin);
-  },
-
-  getLatLngEdges: function () {
-    var tileDegrees = this.dataset.get_baseTileDegrees() / Math.pow(2, this.level);
-    var edges = new LatLngEdges();
-    edges.latMin = (this.dataset.get_baseTileDegrees() / 2 - (this.tileY) * tileDegrees) + this.dataset.get_offsetY();
-    edges.latMax = (this.dataset.get_baseTileDegrees() / 2 - ((this.tileY + 1)) * tileDegrees) + this.dataset.get_offsetY();
-    edges.lngMin = ((this.tileX) * tileDegrees - this.dataset.get_baseTileDegrees() / this.dataset.get_widthFactor()) + this.dataset.get_offsetX();
-    edges.lngMax = (((this.tileX + 1)) * tileDegrees - this.dataset.get_baseTileDegrees() / this.dataset.get_widthFactor()) + this.dataset.get_offsetX();
-    return edges;
-  },
-
-  geoTo3dTan: function (lat, lng) {
-    lng = -lng;
-    var fac1 = this.dataset.get_baseTileDegrees() / 2;
-    var factor = Math.tan(fac1 * Tile.RC);
-    return this.dataset.get_matrix().transform(Vector3d.create(1, (lat / fac1 * factor), (lng / fac1 * factor)));
-  },
-
-  requestImage: function () {
-    this.fitsImage = ss.safeCast(this.dataset.get_wcsImage(), FitsImage);
-    if (this.fitsImage != null) {
-      this.texReady = true;
-      this.downloading = false;
-      this.errored = this.fitsImage.errored;
-      this.requestPending = false;
-      tileCacheRemoveFromQueue(this.get_key(), true);
-      if (useGlVersion2) {
-        this.makeTexture();
-        this.readyToRender = true;
-      }
-      else {
-        this.bmp = this.fitsImage.getBitmap();
-        this.texture2d = this.bmp.getTexture();
-        this.readyToRender = true;
-      }
-    } else {
-      Tile.prototype.requestImage.call(this);
-    }
-  },
-
-  createGeometry: function (renderContext) {
-    if (this.geometryCreated) {
-      return true;
-    }
-    this.geometryCreated = true;
-    for (var i = 0; i < 4; i++) {
-      this._renderTriangleLists[i] = [];
-    }
-    this.globalCenter = this.geoTo3dTan(0, 0);
-    var edges = this.getLatLngEdges();
-    this.topLeft = this.geoTo3dTan(edges.latMin, edges.lngMin).subtract(this.globalCenter);
-    this.bottomRight = this.geoTo3dTan(edges.latMax, edges.lngMax).subtract(this.globalCenter);
-    this.topRight = this.geoTo3dTan(edges.latMin, edges.lngMax).subtract(this.globalCenter);
-    this.bottomLeft = this.geoTo3dTan(edges.latMax, edges.lngMin).subtract(this.globalCenter);
-    var center = Vector3d.midPoint(this.topLeft, this.bottomRight);
-    var leftCenter = Vector3d.midPoint(this.topLeft, this.bottomLeft);
-    var rightCenter = Vector3d.midPoint(this.topRight, this.bottomRight);
-    var topCenter = Vector3d.midPoint(this.topLeft, this.topRight);
-    var bottomCenter = Vector3d.midPoint(this.bottomLeft, this.bottomRight);
-    if (renderContext.gl == null) {
-      this._renderTriangleLists[0].push(RenderTriangle.create(PositionTexture.createPos(this.topLeft, 0, 0), PositionTexture.createPos(leftCenter, 0, 0.5), PositionTexture.createPos(topCenter, 0.5, 0), this.texture, this.level));
-      this._renderTriangleLists[0].push(RenderTriangle.create(PositionTexture.createPos(leftCenter, 0, 0.5), PositionTexture.createPos(center, 0.5, 0.5), PositionTexture.createPos(topCenter, 0.5, 0), this.texture, this.level));
-      this._renderTriangleLists[1].push(RenderTriangle.create(PositionTexture.createPos(topCenter, 0.5, 0), PositionTexture.createPos(rightCenter, 1, 0.5), PositionTexture.createPos(this.topRight, 1, 0), this.texture, this.level));
-      this._renderTriangleLists[1].push(RenderTriangle.create(PositionTexture.createPos(topCenter, 0.5, 0), PositionTexture.createPos(center, 0.5, 0.5), PositionTexture.createPos(rightCenter, 1, 0.5), this.texture, this.level));
-      this._renderTriangleLists[2].push(RenderTriangle.create(PositionTexture.createPos(leftCenter, 0, 0.5), PositionTexture.createPos(bottomCenter, 0.5, 1), PositionTexture.createPos(center, 0.5, 0.5), this.texture, this.level));
-      this._renderTriangleLists[2].push(RenderTriangle.create(PositionTexture.createPos(leftCenter, 0, 0.5), PositionTexture.createPos(this.bottomLeft, 0, 1), PositionTexture.createPos(bottomCenter, 0.5, 1), this.texture, this.level));
-      this._renderTriangleLists[3].push(RenderTriangle.create(PositionTexture.createPos(center, 0.5, 0.5), PositionTexture.createPos(this.bottomRight, 1, 1), PositionTexture.createPos(rightCenter, 1, 0.5), this.texture, this.level));
-      this._renderTriangleLists[3].push(RenderTriangle.create(PositionTexture.createPos(center, 0.5, 0.5), PositionTexture.createPos(bottomCenter, 0.5, 1), PositionTexture.createPos(this.bottomRight, 1, 1), this.texture, this.level));
-      this.readyToRender = true;
-    } else {
-      this._vertexBuffer = tilePrepDevice.createBuffer();
-      tilePrepDevice.bindBuffer(WEBGL.ARRAY_BUFFER, this._vertexBuffer);
-      var f32array = new Float32Array(9 * 5);
-      var buffer = f32array;
-      var index = 0;
-      index = this.addVertex(buffer, index, PositionTexture.createPos(bottomCenter, 0.5, 1));
-      index = this.addVertex(buffer, index, PositionTexture.createPos(this.bottomLeft, 0, 1));
-      index = this.addVertex(buffer, index, PositionTexture.createPos(this.bottomRight, 1, 1));
-      index = this.addVertex(buffer, index, PositionTexture.createPos(center, 0.5, 0.5));
-      index = this.addVertex(buffer, index, PositionTexture.createPos(leftCenter, 0, 0.5));
-      index = this.addVertex(buffer, index, PositionTexture.createPos(rightCenter, 1, 0.5));
-      index = this.addVertex(buffer, index, PositionTexture.createPos(topCenter, 0.5, 0));
-      index = this.addVertex(buffer, index, PositionTexture.createPos(this.topLeft, 0, 0));
-      index = this.addVertex(buffer, index, PositionTexture.createPos(this.topRight, 1, 0));
-      tilePrepDevice.bufferData(WEBGL.ARRAY_BUFFER, f32array, WEBGL.STATIC_DRAW);
-      for (var i = 0; i < 4; i++) {
-        index = 0;
-        this.triangleCount = 2;
-        var ui16array = new Uint16Array(this.triangleCount * 3);
-        var indexArray = ui16array;
-        switch (i) {
-          case 0:
-            indexArray[index++] = 7;
-            indexArray[index++] = 4;
-            indexArray[index++] = 6;
-            indexArray[index++] = 4;
-            indexArray[index++] = 3;
-            indexArray[index++] = 6;
-            break;
-          case 1:
-            indexArray[index++] = 6;
-            indexArray[index++] = 5;
-            indexArray[index++] = 8;
-            indexArray[index++] = 6;
-            indexArray[index++] = 3;
-            indexArray[index++] = 5;
-            break;
-          case 2:
-            indexArray[index++] = 4;
-            indexArray[index++] = 0;
-            indexArray[index++] = 3;
-            indexArray[index++] = 4;
-            indexArray[index++] = 1;
-            indexArray[index++] = 0;
-            break;
-          case 3:
-            indexArray[index++] = 3;
-            indexArray[index++] = 2;
-            indexArray[index++] = 5;
-            indexArray[index++] = 3;
-            indexArray[index++] = 0;
-            indexArray[index++] = 2;
-            break;
-        }
-        this._indexBuffers[i] = tilePrepDevice.createBuffer();
-        tilePrepDevice.bindBuffer(WEBGL.ELEMENT_ARRAY_BUFFER, this._indexBuffers[i]);
-        tilePrepDevice.bufferData(WEBGL.ELEMENT_ARRAY_BUFFER, ui16array, WEBGL.STATIC_DRAW);
-      }
-    }
-    return true;
-  }
-};
-
-registerType("TangentTile", [TangentTile, TangentTile$, Tile]);
-
 // wwtlib.ToastTile
 
 export function ToastTile() {
@@ -30936,7 +29977,7 @@ var BitmapOverlay$ = {
     var $this = this;
 
     try {
-      if (RenderContext.useGl) {
+      if (useGl) {
         this.texture2d = this.get_owner().get_owner().getCachedTexture2d(this._filename$1);
         this._textureReady$1 = true;
       }
@@ -30951,7 +29992,7 @@ var BitmapOverlay$ = {
   },
 
   draw3D: function (renderContext, designTime) {
-    if (RenderContext.useGl) {
+    if (useGl) {
       if (this.texture2d == null) {
         this.initializeTexture();
       }
@@ -31036,7 +30077,7 @@ var TextOverlay$ = {
   },
 
   draw3D: function (renderContext, designTime) {
-    if (RenderContext.useGl) {
+    if (useGl) {
       this.initializeTexture();
       this.initializeGeometry();
       this.updateRotation();
@@ -31165,7 +30206,7 @@ var TextOverlay$ = {
   },
 
   initializeGeometry: function () {
-    if (RenderContext.useGl) {
+    if (useGl) {
       Overlay.prototype.initializeGeometry.call(this);
     }
   }
@@ -31205,7 +30246,7 @@ var ShapeOverlay$ = {
   },
 
   draw3D: function (renderContext, designTime) {
-    if (RenderContext.useGl) {
+    if (useGl) {
       this.initializeGeometry();
       this._sprite$1.draw(renderContext, this.points, this.points.length, null, this._triangleStrip$1, this.get_opacity());
     } else {
@@ -32347,60 +31388,6 @@ var CollectionLoadedEventArgs$ = {
 };
 
 registerType("CollectionLoadedEventArgs", [CollectionLoadedEventArgs, CollectionLoadedEventArgs$, ss.EventArgs]);
-
-// wwtlib.SkyImageTile
-
-export function SkyImageTile(level, x, y, dataset, parent) {
-  this.pixelCenterX = 0;
-  this.pixelCenterY = 0;
-  this.scaleX = 0.01;
-  this.scaleY = 0.01;
-  this.height = 0;
-  this.width = 0;
-  TangentTile.call(this, level, x, y, dataset, parent);
-  this.pixelCenterX = dataset.get_offsetX();
-  this.pixelCenterY = dataset.get_offsetY();
-  this.scaleX = -(this.scaleY = dataset.get_baseTileDegrees());
-  if (dataset.get_bottomsUp()) {
-    this.scaleX = -this.scaleX;
-  }
-  this.sphereCenter = this.geoTo3dTan(0, 0);
-  this.radius = 1.25;
-  this.computeBoundingSphere();
-}
-
-var SkyImageTile$ = {
-  getLatLngEdges: function () {
-    var edges = new LatLngEdges();
-    var wcsImage = ss.safeCast(this.dataset.get_wcsImage(), WcsImage);
-    if (wcsImage != null && RenderContext.useGl) {
-      if (useGlVersion2) {
-        this.width = wcsImage.get_sizeX();
-        this.height = wcsImage.get_sizeY();
-      }
-      else {
-        this.height = this.bmp.height;
-        this.width = this.bmp.width;
-        if (this.bmp.height !== wcsImage.get_sizeY()) {
-          this.pixelCenterY += this.bmp.height - wcsImage.get_sizeY();
-        }
-      }
-    } else if (this.texture != null) {
-      this.height = this.texture.naturalHeight;
-      this.width = this.texture.naturalWidth;
-    } else {
-      this.height = 256;
-      this.width = 256;
-    }
-    edges.latMin = 0 + (this.scaleY * (this.height - this.pixelCenterY));
-    edges.latMax = 0 - (this.scaleY * this.pixelCenterY);
-    edges.lngMin = 0 + (this.scaleX * this.pixelCenterX);
-    edges.lngMax = 0 - (this.scaleX * (this.width - this.pixelCenterX));
-    return edges;
-  }
-};
-
-registerType("SkyImageTile", [SkyImageTile, SkyImageTile$, TangentTile]);
 
 // Statics with nontrivial initialization:
